requirements.txt
streamlit>=1.31.0
openai>=1.0.0
requests>=2.31.0
pydub
numpy>=1.24.0
streamlit-folium>=0.15.0
folium>=0.14.0
geopy
elevenlabs
python-dotenv>=1.0.0
opencv-python>=4.8.0

# Video processing dependencies (specific versions)
moviepy>=1.0.3
decorator==4.4.2
imageio==2.31.1
imageio-ffmpeg==0.4.8
Pillow>=10.0.0
proglog==0.1.10
tqdm>=4.11.2

googlemaps>=4.10.0
imageio[pyav]
av
opencv-python-headless
google-generativeai>=0.3.0

# Add this line to your requirements.txt file or install it manually
streamlit_drawable_canvas>=0.9.2
google-cloud-aiplatform>=1.35.0